prediction_portal/requirement.txt
django
djangorestframework
django-cors-headers
tensorflow
yfinance
scikit-learn
joblib
numpy
djangorestframework-simplejwt 